Refactor packages/app/src/app/pages/Dashboard/Content/CreateNewSandbox/index.tsx
